Move enrichment into sub-package
