+++ server/src/main/java/io/spine/server/event/enrich/SameTransition.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.event;
+package io.spine.server.event.enrich;
 
 import org.checkerframework.checker.nullness.qual.Nullable;
