+++ server/src/main/java/io/spine/server/event/enrich/SupportsFieldConversion.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.event;
+package io.spine.server.event.enrich;
 
 import com.google.common.annotations.VisibleForTesting;
 import io.spine.annotation.Internal;
