+++ server/src/main/java/io/spine/system/server/SystemContext.java
@@ -23,8 +23,8 @@ package io.spine.system.server;
 import io.spine.annotation.Internal;
 import io.spine.server.BoundedContext;
 import io.spine.server.BoundedContextBuilder;
-import io.spine.server.event.Enricher;
 import io.spine.server.event.EventBus;
+import io.spine.server.event.enrich.Enricher;
 
 /**
  * An implementation of {@link BoundedContext} used for the System domain.
