+++ server/src/test/java/io/spine/server/event/enrich/EnricherTest.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.event;
+package io.spine.server.event.enrich;
 
 import io.spine.core.Enrichment;
 import io.spine.core.Event;
@@ -27,6 +27,8 @@ import io.spine.core.EventEnvelope;
 import io.spine.core.Subscribe;
 import io.spine.core.UserId;
 import io.spine.server.BoundedContext;
+import io.spine.server.event.AbstractEventSubscriber;
+import io.spine.server.event.EventBus;
 import io.spine.server.event.given.EventEnricherTestEnv.GivenEvent;
 import io.spine.server.event.given.EventEnricherTestEnv.GivenEventMessage;
 import io.spine.test.event.ProjectCompleted;
@@ -87,7 +89,6 @@ public class EnricherTest {
         boundedContext.close();
     }
 
-    @SuppressWarnings("DuplicateStringLiteralInspection") // Common test case.
     @Test
     @DisplayName("have builder")
     void haveBuilder() {
@@ -232,8 +233,7 @@ public class EnricherTest {
      * <p>This class is a part of assert checking, and as such it is not placed under the test
      * {@linkplain io.spine.server.event.given.EventEnricherTestEnv environment class} .
      */
-    @SuppressWarnings({"OptionalGetWithoutIsPresent", "InstanceVariableNamingConvention",
-                       "ConstantConditions"})
+    @SuppressWarnings({"OptionalGetWithoutIsPresent", "InstanceVariableNamingConvention", "unused"})
     private static class TestEventSubscriber extends AbstractEventSubscriber {
 
         private ProjectCreated.Enrichment projectCreatedEnrichment;
