+++ server/src/test/java/io/spine/server/event/enrich/SupportsFieldConversionTest.java
@@ -18,50 +18,23 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.event;
+package io.spine.server.event.enrich;
 
 import com.google.common.testing.NullPointerTester;
-import com.google.protobuf.Empty;
-import com.google.protobuf.Message;
-import io.spine.core.EventContext;
-import io.spine.server.event.given.EventMessageEnricherTestEnv.Enrichment;
-import io.spine.test.event.ProjectCreated;
-import org.junit.jupiter.api.BeforeEach;
+import com.google.protobuf.StringValue;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static io.spine.testing.DisplayNames.NOT_ACCEPT_NULLS;
-import static org.junit.jupiter.api.Assertions.assertFalse;
 
-/**
- * @author Alexander Litus
- */
-@DisplayName("MessageEnrichment should")
-class MessageEnrichmentTest {
-
-    private MessageEnrichment<ProjectCreated, ProjectCreated.Enrichment, ?> enricher;
-
-    @BeforeEach
-    void setUp() {
-        Enricher enricher = Enrichment.newEventEnricher();
-        this.enricher = MessageEnrichment.create(
-                enricher,
-                ProjectCreated.class,
-                ProjectCreated.Enrichment.class);
-    }
+@DisplayName("SupportsFieldConversion should")
+class SupportsFieldConversionTest {
 
     @Test
     @DisplayName(NOT_ACCEPT_NULLS)
     void passNullToleranceCheck() {
-        new NullPointerTester()
-                .setDefault(Message.class, Empty.getDefaultInstance())
-                .setDefault(EventContext.class, EventContext.getDefaultInstance())
-                .testAllPublicInstanceMethods(enricher);
-    }
-
-    @Test
-    @DisplayName("be inactive when created")
-    void beInactiveWhenCreated() {
-        assertFalse(enricher.isActive());
+        SupportsFieldConversion predicate =
+                SupportsFieldConversion.of(StringValue.class, String.class);
+        new NullPointerTester().testAllPublicInstanceMethods(predicate);
     }
 }
