+++ server/src/test/java/io/spine/server/event/given/EventBusTestEnv.java
@@ -47,11 +47,11 @@ import io.spine.server.aggregate.Apply;
 import io.spine.server.bus.BusFilter;
 import io.spine.server.command.Assign;
 import io.spine.server.event.AbstractEventSubscriber;
-import io.spine.server.event.Enricher;
 import io.spine.server.event.EventBus;
 import io.spine.server.event.EventBusTest;
 import io.spine.server.event.EventDispatcher;
 import io.spine.server.event.EventStreamQuery;
+import io.spine.server.event.enrich.Enricher;
 import io.spine.server.integration.ExternalMessageDispatcher;
 import io.spine.server.tenant.TenantAwareOperation;
 import io.spine.test.event.EBProjectArchived;
