+++ server/src/test/java/io/spine/testdata/TestBoundedContextFactory.java
@@ -23,8 +23,8 @@ package io.spine.testdata;
 import io.spine.server.BoundedContext;
 import io.spine.server.BoundedContextBuilder;
 import io.spine.server.commandbus.CommandBus;
-import io.spine.server.event.Enricher;
 import io.spine.server.event.EventBus;
+import io.spine.server.event.enrich.Enricher;
 import io.spine.server.stand.Stand;
 
 /**
